Station Info cache -> 1.day, remove cache miss event
We were previously caching the station list (not the statuses, just the
list) for a program at 30 minutes. The number of station changes most
likely doesn't happen THAT frequently. Let's extend it to 1 day.

Also, the caching works, so don't really need to put ahoy events for
when cache misses happen. They are just cluttering up the events table.